new features
 * key files support
 * prepared for multiple digests/cyphers support
 * code is now even more generic :-)

+++ src/safemem.cc
@@ -38,7 +38,7 @@ void safe_free(void *p)  {
 }
 
 void fill_random(void *p, size_t size) {
-    RAND_bytes((unsigned char *)p, (size_t)size);
+    RAND_bytes((unsigned char *)p, (int)size);
 }
